src/config/settings.py
"""
Application configuration and settings.
"""

import os
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

# Required settings
SLACK_BOT_TOKEN = os.getenv("SLACK_BOT_TOKEN")
if not SLACK_BOT_TOKEN:
    raise ValueError("SLACK_BOT_TOKEN environment variable is not set")

OPENAI_API_KEY = os.getenv("OPENAI_API_KEY")
if not OPENAI_API_KEY:
    raise ValueError("OPENAI_API_KEY environment variable is not set")

# Optional settings with defaults
HOST = os.getenv("HOST", "0.0.0.0")
PORT = int(os.getenv("PORT", "3000"))
DEBUG = os.getenv("DEBUG", "false").lower() == "true" 