requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
slack-sdk>=3.27.0
openai>=1.12.0
certifi>=2024.2.0
pydantic>=2.6.0